Updated Data File, Closest Location Dataset added
Updated the dataset to repair errors with spurious newlines, and lat &
long being included in the wind and solar lists.
Added the key value pairs for the closest valid location given an ID.
Given an ID value of a latitude and longitude coordinate (in the format
"latlong") you use this as a key, and the value paired, is the closest
location for your shelter.
createClosestJSON.py took approximately 2 hours to create closest.json
on an i5.